% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cmp.R
\name{components}
\alias{components}
\alias{seasonal}
\alias{trendcycle}
\alias{irregular}
\alias{seasonaladj}
\alias{calendaradj}
\alias{calendar}
\alias{y_forecast}
\title{Extract Component from 'RJDemetra' model}
\usage{
seasonal(x, forecast = FALSE)

trendcycle(x, forecast = FALSE)

irregular(x, forecast = FALSE)

seasonaladj(x, forecast = FALSE)

calendaradj(x, forecast = FALSE)

calendar(x, forecast = FALSE)

y_forecast(x)
}
\arguments{
\item{x}{a \code{"SA"} or \code{"jSA"} model.}

\item{forecast}{boolean indicating if the forecast series should be returned.}
}
\description{
Extract Component from 'RJDemetra' model
}
