% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cmp.R
\name{deprecated-ggdemetra}
\alias{deprecated-ggdemetra}
\alias{y_forecast}
\title{Deprecated functions}
\usage{
y_forecast(x)
}
\arguments{
\item{x}{a \code{"SA"} or \code{"jSA"} model.}
}
\description{
Use \code{\link{raw}} with parameter \code{forecast = TRUE} instead of \code{y_forecast}.
}
