% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sa.R
\name{geom_sa}
\alias{geom_sa}
\alias{stat_sa}
\title{Seasonal adjustment time series}
\usage{
geom_sa(mapping = NULL, data = NULL, stat = "sa",
  position = "identity", ..., method = c("x13", "tramoseats"),
  spec = NULL, frequency = NULL, message = TRUE, component = "sa",
  show.legend = NA, inherit.aes = TRUE)

stat_sa(mapping = NULL, data = NULL, geom = "line",
  position = "identity", ..., method = c("x13", "tramoseats"),
  spec = NULL, frequency = NULL, message = TRUE, component = "sa",
  show.legend = NA, inherit.aes = TRUE)
}
\description{
Seasonal adjustment time series
}
