% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_label_arima.R
\name{geom_text_arima}
\alias{geom_text_arima}
\alias{geom_label_arima}
\title{ARIMA model}
\usage{
geom_text_arima(mapping = NULL, data = NULL, stat = "arima",
  position = "identity", ..., method = c("x13", "tramoseats"),
  spec = NULL, frequency = NULL, x_arima = NULL, y_arima = NULL,
  show.legend = NA, inherit.aes = TRUE)

geom_label_arima(mapping = NULL, data = NULL, stat = "arima",
  position = "identity", ..., method = c("x13", "tramoseats"),
  spec = NULL, frequency = NULL, message = TRUE, x_arima = NULL,
  y_arima = NULL, show.legend = NA, inherit.aes = TRUE)
}
\description{
\code{geom_text_arima()} adds the ARIMA models directly to the plot. \code{geom_label_arima()} draws a rectangle behind the ARIMA model, making it easier to read.
}
