% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_label_outliers.R
\name{geom_text_outlier}
\alias{geom_text_outlier}
\alias{geom_label_outlier}
\alias{geom_text_repel_outlier}
\alias{geom_label_repel_outlier}
\title{Outliers texts}
\usage{
geom_text_outlier(mapping = NULL, data = NULL, stat = "outlier",
  position = "identity", ..., method = c("x13", "tramoseats"),
  spec = NULL, frequency = NULL, message = TRUE, first_date = NULL,
  last_date = NULL, coefficients = FALSE, digits = 1,
  show.legend = NA, inherit.aes = TRUE)

geom_label_outlier(mapping = NULL, data = NULL, stat = "outlier",
  position = "identity", ..., method = c("x13", "tramoseats"),
  spec = NULL, frequency = NULL, message = TRUE, first_date = NULL,
  last_date = NULL, coefficients = FALSE, digits = 1,
  show.legend = NA, inherit.aes = TRUE)

geom_text_repel_outlier(mapping = NULL, data = NULL,
  stat = "outlier", position = "identity", ..., method = c("x13",
  "tramoseats"), spec = NULL, frequency = NULL, message = TRUE,
  first_date = NULL, last_date = NULL, coefficients = FALSE,
  digits = 1, show.legend = NA, inherit.aes = TRUE)

geom_label_repel_outlier(mapping = NULL, data = NULL,
  stat = "outlier", position = "identity", ..., method = c("x13",
  "tramoseats"), spec = NULL, frequency = NULL, message = TRUE,
  first_date = NULL, last_date = NULL, coefficients = FALSE,
  digits = 1, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \link[ggplot2:aes]{aes()} or
\link[ggplot2:aes_]{aes_()}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{A \code{data.frame} that contains the data used for the seasonal adjustment. By default,}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \link[ggplot2:layer]{layer()}. They may be parameters of
\link[ggplot2:geom_text]{geom_text()}, \link[ggplot2:geom_label]{geom_label()},
\link[ggrepel:geom_text_repel]{geom_text_repel()} or \link[ggrepel:geom_label_repel]{geom_label_repel()}.}

\item{method}{the method used for the seasonal adjustment. \code{"x13"} (by default) for the X-13ARIMA method and \code{"tramoseats"} for TRAMO-SEATS.}

\item{spec}{the specification used for the seasonal adjustment.
See \link[RJDemetra:x13]{x13()} or \link[RJDemetra:tramoseats]{tramoseats()}.}

\item{frequency}{the frequency of the time series. By default (\code{frequency = NULL}),
the frequency is computed automatically.}

\item{message}{a \code{boolean} indicating if a message is printed with the frequency used.}

\item{first_date}{A numeric specifying the first date from which the outliers are plotted.
By default (\code{first_date = NULL}) the outliers are plotted from the
beginning of the time series.}

\item{last_date}{A numeric specifying the first date from which the outliers are plotted.
By default (\code{first_date = NULL}) the outliers are plotted until the
end of the time series.}

\item{coefficients}{boolean indicating if the estimates coefficients are printed.
By default \code{coefficients = FALSE}.}

\item{digits}{integer indicating the number of decimal places to be used for numeric diagnostics. By default \code{digits = 1}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them.}
}
\description{
\code{geom_text_outliers()} adds directly to the plot the outliers used in the pre-adjustment process of the seasonal adjustment. \code{geom_label_outliers()} draws a rectangle behind the names of the outliers, making them easier to read. In \code{geom_text_repel_outliers()} and \code{geom_label_repel_outliers()}, text labels repel away from each other and away from the data points (see \link[ggrepel:geom_label_repel]{geom_label_repel()}).
}
