% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_ggplot.R
\name{init_ggplot}
\alias{init_ggplot}
\title{Initialise 'ggplot2' with 'SA' model}
\usage{
init_ggplot(x, ...)
}
\arguments{
\item{x}{A \code{"SA"} or \code{"jsA"} model created with 'RJDemetra'.}

\item{...}{Other parameters passes to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}
}
\description{
Initialise 'ggplot2' with 'SA' model
}
\examples{
mod <- RJDemetra::x13(ipi_c_eu[, "FR"])
init_ggplot(mod) +
    geom_line(color =  "#F0B400") +
    geom_sa(component = "sa", color = "#155692")
}
