% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts2df.R
\name{ts2df}
\alias{ts2df}
\title{Convert 'ts' object to 'data.frame'}
\usage{
ts2df(x)
}
\arguments{
\item{x}{a \code{ts} or \code{mts} object.}
}
\value{
a \code{data.frame} object.
}
\description{
Function to a \code{ts} or \code{mts} object to a \code{data.frame} that can be directly used in the plot functions.
}
\examples{
# To get the ipi_c_eu_df object:
ts2df(ipi_c_eu)
}
