% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_label_diagnostics.R,
%   R/geom_label_outliers.R
\name{geom_text_outlier}
\alias{geom_text_outlier}
\alias{geom_label_outlier}
\alias{geom_text_repel_outlier}
\alias{geom_label_repel_outlier}
\title{Seasonal adjustment time series}
\usage{
geom_text_outlier(mapping = NULL, data = NULL, stat = "outlier",
  position = "identity", ..., method = c("x13", "tramoseats"),
  spec = NULL, frequency = NULL, message = TRUE,
  coefficients = FALSE, digits = 1, show.legend = NA,
  inherit.aes = TRUE)

geom_label_outlier(mapping = NULL, data = NULL, stat = "outlier",
  position = "identity", ..., method = c("x13", "tramoseats"),
  spec = NULL, frequency = NULL, message = TRUE,
  coefficients = FALSE, digits = 1, show.legend = NA,
  inherit.aes = TRUE)

geom_text_repel_outlier(mapping = NULL, data = NULL,
  stat = "outlier", position = "identity", ..., method = c("x13",
  "tramoseats"), spec = NULL, frequency = NULL, message = TRUE,
  coefficients = FALSE, digits = 1, show.legend = NA,
  inherit.aes = TRUE)

geom_label_repel_outlier(mapping = NULL, data = NULL,
  stat = "outlier", position = "identity", ..., method = c("x13",
  "tramoseats"), spec = NULL, frequency = NULL, message = TRUE,
  coefficients = FALSE, digits = 1, show.legend = NA,
  inherit.aes = TRUE)

geom_text_outlier(mapping = NULL, data = NULL, stat = "outlier",
  position = "identity", ..., method = c("x13", "tramoseats"),
  spec = NULL, frequency = NULL, message = TRUE,
  coefficients = FALSE, digits = 1, show.legend = NA,
  inherit.aes = TRUE)

geom_label_outlier(mapping = NULL, data = NULL, stat = "outlier",
  position = "identity", ..., method = c("x13", "tramoseats"),
  spec = NULL, frequency = NULL, message = TRUE,
  coefficients = FALSE, digits = 1, show.legend = NA,
  inherit.aes = TRUE)

geom_text_repel_outlier(mapping = NULL, data = NULL,
  stat = "outlier", position = "identity", ..., method = c("x13",
  "tramoseats"), spec = NULL, frequency = NULL, message = TRUE,
  coefficients = FALSE, digits = 1, show.legend = NA,
  inherit.aes = TRUE)

geom_label_repel_outlier(mapping = NULL, data = NULL,
  stat = "outlier", position = "identity", ..., method = c("x13",
  "tramoseats"), spec = NULL, frequency = NULL, message = TRUE,
  coefficients = FALSE, digits = 1, show.legend = NA,
  inherit.aes = TRUE)
}
\description{
Seasonal adjustment time series

Seasonal adjustment time series
}
